Move Label and EditBox into their own modules
