+++ src/widget/editbox.rs
@@ -15,76 +15,6 @@ use crate::macros::Widget;
 use crate::{Align, AlignHints, CoreData, Layout, Widget, WidgetCore};
 use kas::geom::Rect;
 
-/// A simple text label
-#[widget]
-#[handler]
-#[derive(Clone, Default, Debug, Widget)]
-pub struct Label {
-    #[core]
-    core: CoreData,
-    align: (Align, Align),
-    text: String,
-}
-
-impl Layout for Label {
-    fn size_rules(&mut self, size_handle: &mut dyn SizeHandle, axis: AxisInfo) -> SizeRules {
-        let rules = size_handle.text_bound(&self.text, TextClass::Label, axis);
-        if axis.is_horizontal() {
-            self.core_data_mut().rect.size.0 = rules.ideal_size();
-        } else {
-            self.core_data_mut().rect.size.1 = rules.ideal_size();
-        }
-        rules
-    }
-
-    fn set_rect(&mut self, _size_handle: &mut dyn SizeHandle, rect: Rect, align: AlignHints) {
-        self.align = (
-            align.horiz.unwrap_or(Align::Begin),
-            align.vert.unwrap_or(Align::Centre),
-        );
-        self.core_data_mut().rect = rect;
-    }
-
-    fn draw(&self, draw_handle: &mut dyn DrawHandle, _: &ManagerState) {
-        draw_handle.text(self.core.rect, &self.text, TextClass::Label, self.align);
-    }
-}
-
-impl Label {
-    /// Construct a new, empty instance
-    pub fn new<T: ToString>(text: T) -> Self {
-        Label {
-            core: Default::default(),
-            align: Default::default(),
-            text: text.to_string(),
-        }
-    }
-}
-
-impl<T> From<T> for Label
-where
-    String: From<T>,
-{
-    fn from(text: T) -> Self {
-        Label {
-            core: Default::default(),
-            align: Default::default(),
-            text: String::from(text),
-        }
-    }
-}
-
-impl HasText for Label {
-    fn get_text(&self) -> &str {
-        &self.text
-    }
-
-    fn set_string(&mut self, mgr: &mut Manager, text: String) {
-        self.text = text;
-        mgr.redraw(self.id());
-    }
-}
-
 #[derive(Clone, Debug, PartialEq)]
 enum LastEdit {
     None,
