+++ src/widget/mod.rs
@@ -11,21 +11,23 @@
 mod button;
 mod checkbox;
 mod dialog;
+mod editbox;
 mod filler;
+mod label;
 mod list;
 mod radiobox;
 mod scroll;
 mod scrollbar;
-mod text;
 mod window;
 
 pub use button::TextButton;
 pub use checkbox::{CheckBox, CheckBoxBare};
 pub use dialog::MessageBox;
+pub use editbox::EditBox;
 pub use filler::Filler;
+pub use label::Label;
 pub use list::{BoxColumn, BoxList, BoxRow, Column, List, Row};
 pub use radiobox::{RadioBox, RadioBoxBare};
 pub use scroll::ScrollRegion;
 pub use scrollbar::ScrollBar;
-pub use text::{EditBox, Label};
 pub use window::Window;
